+++ Server/server/Map/Case.hpp
@@ -7,18 +7,18 @@
 class Case : public WorldObject
 {
 public:
-	Case(uint16, uint16, uint16);
+	Case(uint32, uint16, uint16);
 	~Case();
 	void SetBlock(bool);
     bool CanBeWalk();
     bool IsBlocking() const;
-    uint16 GetID() const;
+    uint32 GetID() const;
     void AddDynamicOject(DynamicObject*);
     void UnitEnterInCase(Unit*);
 
 private:
 	bool m_Block;
-	uint16 m_ID;
+	uint32 m_ID;
     std::vector<DynamicObject*> m_DynamicObjectList;
 };
